requirements.txt
streamlit==1.40.0
ultralytics
opencv-python-headless
Pillow
numpy
streamlit-drawable-canvas
pathlib
